skip mismatch site with sequence N in reads
