+++ src/giremi/script/get_aei_short.py
@@ -182,7 +182,7 @@ def alu_df_aei(alu, variables):
 ####################
 
 
-if __name__ == '__main__':
+def parse_args():
     parser = argparse.ArgumentParser(
         description='Calculate AEI from bam file, both in read level and Alu level'
     )
@@ -253,7 +253,11 @@ if __name__ == '__main__':
         default = 1
     )
     args = parser.parse_args()
+    return args
 
+
+def main():
+    args = parse_args()
     variables = {
         'bam_file': args.bam_file,
         'outprefix': args.output_prefix,
@@ -292,5 +296,9 @@ if __name__ == '__main__':
         )
     else:
         pass
+
+
+if __name__ == '__main__':
+    main()
 
 ####################
